Har fjernet Main og gjort actionMenu til main
